Renamed two files for better clarity
