+++ package.json
@@ -1,73 +1,76 @@
 {
-  
+  
   "name": "api-backend",
-  
+  
   "version": "2022.2",
-  
+  
   "description": "A starter express api backend that can also serve a React (CRA) client",
-  
+  
   "main": "index.js",
-  
+  
   "scripts": {
-  
-    
+  
+    
   "start": "node api/app.js",
-  
-    
+  
+    
   "dev": "nodemon -r dotenv/config api/app.js",
-  
-    
+  
+    
   "test": "echo \"Error: no test specified\" && exit 1",
-  
-    
+  
+    
   "build": "cd client && npm install && npm run build"
-  
+  
   },
-  
+  
   "author": "Edgardo Molina",
-  
+  
   "license": "MIT",
-  
+  
   "dependencies": {
-  
-    
+  
+    
   "express": "^4.18.2",
-  
-    
+  
+    
+  "flowbite": "^1.5.3",
+  
+    
   "morgan": "^1.10.0",
-  
-    
+  
+    
   "pg": "^8.8.0",
-  
-    
+  
+    
   "pg-hstore": "^2.3.4",
-  
-    
+  
+    
   "sequelize": "^6.25.2"
-  
+  
   },
-  
+  
   "devDependencies": {
-  
-    
+  
+    
   "dotenv": "^8.6.0",
-  
-    
+  
+    
   "nodemon": "^2.0.20"
-  
+  
   },
-  
+  
   "nodemonConfig": {
-  
-    
+  
+    
   "ignore": [
-  
-    
-    
+  
+    
+    
   "client/"
-  
-    
+  
+    
   ]
-  
+  
   }
 }
